Revert "Train loop subpackage"
