+++ tests/test_torchtrain/test_tl_components/test_callback_handler.py
@@ -1,7 +1,7 @@
 import unittest
 
 from tests.utils import *
-from aitoolbox.torchtrain.train_loop.components.callback_handler import CallbacksHandler, BasicCallbacksHandler
+from aitoolbox.torchtrain.tl_components.callback_handler import CallbacksHandler, BasicCallbacksHandler
 from aitoolbox.torchtrain.callbacks.abstract import AbstractCallback
 from aitoolbox.torchtrain.train_loop import TrainLoop
