+++ tests/test_torchtrain/test_tl_components/test_message_passing.py
@@ -1,6 +1,6 @@
 import unittest
 
-from aitoolbox.torchtrain.train_loop.components import message_passing as msg_pass
+from aitoolbox.torchtrain.tl_components import message_passing as msg_pass
 
 
 class TestMessageService(unittest.TestCase):
