+++ tests/test_torchtrain/test_tl_components/test_pred_collate_fns.py
@@ -1,7 +1,8 @@
 import unittest
 import numpy as np
+import torch
 
-from aitoolbox.torchtrain.train_loop.components.pred_collate_fns import *
+from aitoolbox.torchtrain.tl_components.pred_collate_fns import *
 
 
 class TestBatchPredCollateFns(unittest.TestCase):
