+++ tests_gpu/test_multi_gpu/test_core_pytorch_compare/test_dp/test_imdb_rnn_lstm_experiment_track.py
@@ -10,10 +10,9 @@ import torch.nn as nn
 import torchtext
 import torchtext.data
 
-from aitoolbox import TTModel, TTDataParallel, \
+from aitoolbox import TrainLoopCheckpointEndSave, TTModel, TTDataParallel, \
     ModelPerformanceEvaluation, ModelPerformancePrintReport, \
     ModelTrainHistoryPlot, ModelTrainHistoryFileWriter, BinaryClassificationResultPackage
-from aitoolbox.torchtrain.train_loop import TrainLoopCheckpointEndSave
 
 THIS_DIR = os.path.dirname(os.path.abspath(__file__))
